refactor: share telefone normalization
